Refactor folder structures, add simple syncing test
