+++ package.json
@@ -20,7 +20,7 @@
     "eslint": "^7.26.0",
     "ganache-cli": "^6.12.2",
     "husky": "^7.0.2",
-    "mangata-sdk": "^0.2.1",
+    "mangata-sdk": "https://github.com/mangata-finance/mangata-sdk",
     "package.json": "^2.0.1",
     "prettier": "^2.3.0",
     "selenium-webdriver": "3.6.0",
@@ -39,7 +39,7 @@
     "test": "jest",
     "test-parallel": "rm nonce.db ; rm sudo.lock; jest --verbose --ci --group=parallel",
     "test-playground-pararell": "rm nonce.db ; rm sudo.lock; jest test/exploratory/parallelTesting/transfer/* --verbose",
-    "test-cluster": "npx ts-node ./test/cluster/index.ts",
+    "test-cluster": "rm nonce.db; rm sudo.lock; jest --verbose --group=cluster",
     "test-sequential": "rm nonce.db ; rm sudo.lock; jest --verbose --runInBand --group=sequential",
     "test-sequential-e2e": "rm nonce.db ; rm sudo.lock; jest --verbose --runInBand --ci --group=critical ",
     "test-ci": "rm nonce.db ; rm sudo.lock; jest --runInBand --verbose --ci --group=ci ",
@@ -53,6 +53,7 @@
     "@babel/preset-env": "^7.13.10",
     "@babel/preset-typescript": "^7.13.0",
     "@types/jest": "^26.0.23",
+    "@types/lodash": "^4.14.173",
     "babel-jest": "^26.6.3",
     "eslint-config-mangata": "https://github.com/mangata-finance/eslint-config-mangata",
     "jest": "^26.6.3",
@@ -60,6 +61,7 @@
     "jest-html-reporters": "^2.1.6",
     "jest-junit": "^12.2.0",
     "jest-runner-groups": "^2.1.0",
+    "lodash": "^4.17.21",
     "mangata-bridge": "https://github.com/mangata-finance/mangata-bridge",
     "mangata-node": "https://github.com/mangata-finance/mangata-node"
   },
